Remove usage of Google Cloud APIs
I was reading the code and there were too much references to Google
Cloud APIs. I've seen that you already deleted unwanted AppEngine code
so I hope it's also in the direction you want to take this code.

I was also thinking about trying to add support for GOPROXY, I'm working
in an intranet with no internet connection and I'm trying to make sort
of a mirror of godocs.io. What do you think, would it be something that
fits this fork?