gddo-server: Remove importerCount limit in Etag
Adnan Maolood, also unsure what this limit on `importerCount` is for,
suggested that it could be removed.

Perhaps it was there so that frequently imported packages wouldn't
constantly have their Etags invalidated, but I don't know why "8" was
chosen as a limit, and can't find relevant information in the Git logs.

Since we're not clear on why the limit was introduced, remove it.