gddo-server/fetch: Try parent import paths on proxy.ErrNotFound
Previously, fetch would fail when presented with the import path of a
subpackage (e.g. example.org/module/package). This meant that linking to
the documentation for versions not present in the database would lead to
a 404 error. We now try parent import paths until one succeeds or we
receive a failure.

An alternative approach which was considered was to separate the module
path from the import path in package URLs. The resulting URLs would look
something like https://godocs.io/example.org/module@version/package.
The placement of @version would allow determining the module path
(example.org/module) without unnecessary requests to the module proxy.
However, this approach was abandoned because it would either: (a) break
existing URLs, (b) not be able to handle existing URLs, or (c)
complicate the request parsing logic and internals of the server. For
these reasons a simple loop seems like the better solution.

Resolves #26.