internal/stdlib: Treat stdlib as multiple modules
Instead of treating the stdlib as a single module named "std", treat
each top-level directory as a module. This treats 'archive' as a module
and 'archive/tar' as a subpackage of the 'archive' module. This change
allows us to remove some stdlib-specific exceptions throughout the
codebase, and will also enable us to link to the source code of stdlib
packages (#28).

The existing 'std' module in the database can be safely deleted once
this is deployed.