gddo-server: Implement ability to choose platform
Implement the ability to choose the platform the documentation is
displayed in. The database structure was simplified: the documentation
and imports tables were merged into the packages table. Templates were
also modified to keep the requested platform in all the necessary links,
so that clicking a link won't reset your preference.

Resolves #8