Check for X-Forwarded-Proto in getRootURL
Currently, 'getRootURL' is configured to only use 'https' for the proto
when the request is using TLS. This is not actually achievable when
using a reverse proxy for TLS termination, so the root URL never has
'https' for the proto.

This commit adds a checks for a protocol hint in the 'X-Forwarded-Proto'
proxy header. If that header is found, the value will be used for the
protocol.