internal/server: Support serving modules from GOMODCACHE
This patch implements support for serving modules from GOMODCACHE. This
is a read-only operation: users must run 'go mod download' manually to
download modules to the module cache. This is probably best used for
viewing documentation locally.

This required some changes to the gddo-server command:

- gddo-server can now operate without a PostgreSQL database, and will
only attempt to connect to one if the --db flag is provided.
- gddo-server will use the modcache by default unless the --proxy
flag is specified.
- The module cache directory can be set with the new --modcache flag. If
unset, it defaults to the output of "go env GOMODCACHE".
- The module loading function was refactored to support partial loading
of module data, similar to golang.org/x/tools/go/packages' LoadMode.

Modules are read from $GOMODCACHE/cache/download, which is formatted
similar to the GOPROXY. However, module source code is read from the
$GOMODCACHE directory which contains decompressed source code to avoid
decompressing the module zip file on every request.

Since standard library modules aren't present in the GOMODCACHE
directory, they are read from $GOROOT/src. This allows for fully offline
browsing of documentation. However, only the currently installed version
of the standard library can be viewed.