internal/database: Fix misleading package versions
When fetching a package with GetPackage, the database wrongly sets the
package's Versions to the list of versions in the module. This is wrong
as different module versions are not guaranteed to contain the same
packages. This lead to invalid links and made it harder for users to
determine which versions a package actually contained.

Fix by only listing the versions present in the database. This is less
than ideal as this list might be incomplete, but it is better to be
incomplete than to be misleading.

Also add a note to the versions templates that the list presented may be
incomplete, with a link to the full list of versions in the module.

Fixes: https://todo.sr.ht/~sircmpwn/gddo/37