internal/server: Add release tags to build contexts
Release tags (e.g. 'go1.8') were mistakenly not enabled by gddo. The
documentation for go/build.Context says: "it is usually an error to
customize ReleaseTags, which defaults to the list of Go releases the
current release is compatible with". However, it seems that this is
referring to the default build context, so we'll have to provide our own
set of ReleaseTags when we create our own build context. Here we just
use the default set from build.Default.ReleaseTags.