internal: Add Package.Reference field
Add a Reference field to distinguish between versions and references
(e.g. git tags). This is useful for standard library packages, where the
reference and version are not equal. To implement this, a column was
added to the database and module sources were modified to fill in the
Reference field.

Also take the opportunity to clean up the templating code in
internal/server and fix a few issues with source links.