internal/server: Don't load packages in the background
Previously, we were wrongly loading packages from the database using the
background context. This meant that receiving a lot of requests resulted
in a lot of pending database operations which would not be cancelled
when the request is cancelled. This allowed a malicious actor or poorly
written crawler to DoS the site. Instead, use the request context to
load packages from the database so that operations are properly
cancelled.